src/requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
django-crispy-forms==1.9.1
django-environ==0.4.5
django-nested-admin==3.3.0
Django==3.0.6
djangorestframework-jwt==1.11.0
djangorestframework==3.11.0
idna==2.9
mysqlclient==1.4.6
Pillow==7.1.2
PyJWT==1.7.1
pytz==2020.1
requests==2.23.0
sqlparse==0.3.1
urllib3==1.25.9
